start modularizing integration tests
