+++ tests/integration/__init__.py
@@ -1,0 +1,6 @@
+# -*- mode: python; coding: utf-8 -*-
+#
+# Copyright (C) 2025 Benjamin Thomas Schwertfeger
+# All rights reserved.
+# https://github.com/btschwertfeger
+#
