+++ tests/integration/kraken_exchange/conftest.py
@@ -1,0 +1,20 @@
+# -*- mode: python; coding: utf-8 -*-
+#
+# Copyright (C) 2025 Benjamin Thomas Schwertfeger
+# All rights reserved.
+# https://github.com/btschwertfeger
+#
+
+import pytest
+
+from .kraken_exchange_api import KrakenExchangeAPIConfig
+
+
+@pytest.fixture(scope="session")
+def kraken_config_xbtusd() -> KrakenExchangeAPIConfig:
+    return KrakenExchangeAPIConfig(
+        base_currency="XXBT",
+        quote_currency="ZUSD",
+        pair="XBTUSD",
+        ws_symbol="BTC/USD",
+    )
