+++ tests/integration/kraken_exchange/test_kraken_swing.py
@@ -19,7 +19,8 @@ from infinity_grid.models.configuration import (
     NotificationConfigDTO,
 )
 
-from .helper import get_kraken_instance
+from .helper import KrakenTestManager
+from .kraken_exchange_api import KrakenExchangeAPIConfig
 
 LOG = logging.getLogger(__name__)
 
@@ -57,6 +58,7 @@ async def test_kraken_swing(
     kraken_swing_bot_config: BotConfigDTO,
     notification_config: NotificationConfigDTO,
     db_config: DBConfigDTO,
+    kraken_config_xbtusd: KrakenExchangeAPIConfig,
 ) -> None:
     """
     Integration test for the SWING strategy using pre-generated websocket
@@ -65,16 +67,18 @@ async def test_kraken_swing(
     LOG.info("******* Starting SWING integration test *******")
     caplog.set_level(logging.DEBUG)
 
-    # Create engine using mocked Kraken API
-    engine = await get_kraken_instance(
+    tm = KrakenTestManager(
         bot_config=kraken_swing_bot_config,
         notification_config=notification_config,
         db_config=db_config,
+        kraken_config=kraken_config_xbtusd,
     )
-    state_machine = engine._BotEngine__state_machine
-    strategy = engine._BotEngine__strategy
-    ws_client = strategy._GridHODLStrategy__ws_client
-    api = engine._BotEngine__strategy._GridHODLStrategy__ws_client.__websocket_service
+    await tm.initialize_engine()
+
+    state_machine = tm.state_machine
+    strategy = tm.strategy
+    ws_client = tm.ws_client
+    api = tm.ws_client.__websocket_service
 
     # ==========================================================================
     # During the following processing, the following steps are done:
@@ -82,20 +86,21 @@ async def test_kraken_swing(
     # 2. The order manager checks the price range
     # 3. The order manager checks for n open buy orders
     # 4. The order manager places new orders
-    await ws_client.on_message(
-        {
-            "channel": "executions",
-            "type": "snapshot",
-            "data": [{"exec_type": "canceled", "order_id": "txid0"}],
-        },
-    )
-    assert state_machine.state == States.INITIALIZING
-    assert strategy._ready_to_trade is False
+    await tm.trigger_prepare_for_trading()
+    # await ws_client.on_message(
+    #     {
+    #         "channel": "executions",
+    #         "type": "snapshot",
+    #         "data": [{"exec_type": "canceled", "order_id": "txid0"}],
+    #     },
+    # )
+    # assert state_machine.state == States.INITIALIZING
+    # assert strategy._ready_to_trade is False
 
-    await api.on_ticker_update(callback=ws_client.on_message, last=50000.0)
-    assert strategy._ticker == 50000.0
-    assert state_machine.state == States.RUNNING
-    assert strategy._ready_to_trade is True
+    # await api.on_ticker_update(callback=ws_client.on_message, last=50000.0)
+    # assert strategy._ticker == 50000.0
+    # assert state_machine.state == States.RUNNING
+    # assert strategy._ready_to_trade is True
 
     # ==========================================================================
     # 1. PLACEMENT OF INITIAL N BUY ORDERS
@@ -104,20 +109,25 @@ async def test_kraken_swing(
     # finally saved those results into the local orderbook table.
     # The SWING strategy additionally starts selling the existing base currency
     # at defined intervals.
-    LOG.info("******* Check placement of initial buy orders *******")
+    # LOG.info("******* Check placement of initial buy orders *******")
+    await tm.trigger_initial_n_buy_orders(
+        prices=(49504.9, 49014.7, 48529.4, 48048.9, 47573.1, 51005.0),
+        volumes=(0.00202, 0.0020402, 0.0020606, 0.00208121, 0.00210202, 0.00197044),
+        sides=("buy", "buy", "buy", "buy", "buy", "sell"),
+    )
 
-    for order, price, volume, side in zip(
-        strategy._orderbook_table.get_orders().all(),
-        (49504.9, 49014.7, 48529.4, 48048.9, 47573.1, 51005.0),
-        (0.00202, 0.0020402, 0.0020606, 0.00208121, 0.00210202, 0.00197044),
-        ["buy"] * 5 + ["sell"],
-        strict=True,
-    ):
-        assert order.price == price
-        assert order.volume == volume
-        assert order.side == side
-        assert order.symbol == "XBTUSD"
-        assert order.userref == strategy._config.userref
+    # for order, price, volume, side in zip(
+    #     strategy._orderbook_table.get_orders().all(),
+    #     (49504.9, 49014.7, 48529.4, 48048.9, 47573.1, 51005.0),
+    #     (0.00202, 0.0020402, 0.0020606, 0.00208121, 0.00210202, 0.00197044),
+    #     ["buy"] * 5 + ["sell"],
+    #     strict=True,
+    # ):
+    #     assert order.price == price
+    #     assert order.volume == volume
+    #     assert order.side == side
+    #     assert order.symbol == "XBTUSD"
+    #     assert order.userref == strategy._config.userref
 
     # ==========================================================================
     # 2. RAPID PRICE DROP - FILLING ALL BUY ORDERS + CREATING SELL ORDERS
@@ -240,6 +250,7 @@ async def test_kraken_swing_unfilled_surplus(
     kraken_swing_bot_config: BotConfigDTO,
     notification_config: NotificationConfigDTO,
     db_config: DBConfigDTO,
+    kraken_config_xbtusd: KrakenExchangeAPIConfig,
 ) -> None:
     """
     Integration test for the SWING strategy using pre-generated websocket
@@ -253,17 +264,19 @@ async def test_kraken_swing_unfilled_surplus(
     LOG.info("******* Starting SWING unfilled surplus integration test *******")
     caplog.set_level(logging.INFO)
 
-    # Create engine using mocked Kraken API
-    engine = await get_kraken_instance(
+    tm = KrakenTestManager(
         bot_config=kraken_swing_bot_config,
         notification_config=notification_config,
         db_config=db_config,
+        kraken_config=kraken_config_xbtusd,
     )
-    state_machine = engine._BotEngine__state_machine
-    strategy = engine._BotEngine__strategy
-    ws_client = strategy._GridHODLStrategy__ws_client
-    rest_api = strategy._rest_api
-    api = engine._BotEngine__strategy._GridHODLStrategy__ws_client.__websocket_service
+    await tm.initialize_engine()
+
+    state_machine = tm.state_machine
+    strategy = tm.strategy
+    ws_client = tm.ws_client
+    rest_api = tm.rest_api
+    api = tm.ws_client.__websocket_service
 
     # ==========================================================================
     # During the following processing, the following steps are done:
@@ -271,43 +284,50 @@ async def test_kraken_swing_unfilled_surplus(
     # 2. The order manager checks the price range
     # 3. The order manager checks for n open buy orders
     # 4. The order manager places new orders
-    await ws_client.on_message(
-        {
-            "channel": "executions",
-            "type": "snapshot",
-            "data": [{"exec_type": "canceled", "order_id": "txid0"}],
-        },
-    )
-    assert state_machine.state == States.INITIALIZING
-    assert strategy._ready_to_trade is False
+    await tm.trigger_prepare_for_trading()
+    # await ws_client.on_message(
+    #     {
+    #         "channel": "executions",
+    #         "type": "snapshot",
+    #         "data": [{"exec_type": "canceled", "order_id": "txid0"}],
+    #     },
+    # )
+    # assert state_machine.state == States.INITIALIZING
+    # assert strategy._ready_to_trade is False
 
-    await api.on_ticker_update(callback=ws_client.on_message, last=50000.0)
-    assert strategy._ticker == 50000.0
-    assert state_machine.state == States.RUNNING
-    assert strategy._ready_to_trade is True
+    # await api.on_ticker_update(callback=ws_client.on_message, last=50000.0)
+    # assert strategy._ticker == 50000.0
+    # assert state_machine.state == States.RUNNING
+    # assert strategy._ready_to_trade is True
 
     # ==========================================================================
     # 1. PLACEMENT OF INITIAL N BUY ORDERS
     # After both fake-websocket channels are connected, the algorithm went
     # through its full setup and placed orders against the fake Kraken API and
     # finally saved those results into the local orderbook table.
-    LOG.info("******* Check placement of initial buy orders *******")
+    # LOG.info("******* Check placement of initial buy orders *******")
 
     # Check if the five initial buy orders are placed with the expected price
     # and volume. Note that the interval is not exactly 0.01 due to the fee
     # which is taken into account.
-    for order, price, volume, side in zip(
-        strategy._orderbook_table.get_orders().all(),
-        (49504.9, 49014.7, 48529.4, 48048.9, 47573.1, 51005.0),
-        (0.00202, 0.0020402, 0.0020606, 0.00208121, 0.00210202, 0.00197044),
-        ["buy"] * 5 + ["sell"],
-        strict=True,
-    ):
-        assert order.price == price
-        assert order.volume == volume
-        assert order.side == side
-        assert order.symbol == "XBTUSD"
-        assert order.userref == strategy._config.userref
+    await tm.trigger_initial_n_buy_orders(
+        prices=(49504.9, 49014.7, 48529.4, 48048.9, 47573.1, 51005.0),
+        volumes=(0.00202, 0.0020402, 0.0020606, 0.00208121, 0.00210202, 0.00197044),
+        sides=("buy", "buy", "buy", "buy", "buy", "sell"),
+    )
+
+    # for order, price, volume, side in zip(
+    #     strategy._orderbook_table.get_orders().all(),
+    #     (49504.9, 49014.7, 48529.4, 48048.9, 47573.1, 51005.0),
+    #     (0.00202, 0.0020402, 0.0020606, 0.00208121, 0.00210202, 0.00197044),
+    #     ["buy"] * 5 + ["sell"],
+    #     strict=True,
+    # ):
+    #     assert order.price == price
+    #     assert order.volume == volume
+    #     assert order.side == side
+    #     assert order.symbol == "XBTUSD"
+    #     assert order.userref == strategy._config.userref
 
     balances = api.get_balances()
     assert float(balances["XXBT"]["balance"]) == pytest.approx(99.99802956)
